refactor: extract is_admin() to shared utility
Move the is_admin() function from admin_commands.py and user_commands.py
to a new shared utility module typer_bot/utils/permissions.py to eliminate
code duplication.

Changes:
- Create typer_bot/utils/permissions.py with is_admin() function
- Export is_admin from typer_bot/utils/__init__.py
- Update admin_commands.py to import from utils
- Update user_commands.py to import from utils and remove _is_admin method

Follows DRY principle - admin role checking logic now exists in one place.
