+++ typer_bot/utils/__init__.py
@@ -1,5 +1,6 @@
 """Utility functions and helpers."""
 
+from .permissions import is_admin
 from .prediction_parser import (
     ascii_username,
     format_standings,
@@ -11,6 +12,7 @@ from .timezone import APP_TZ, format_for_discord, now, parse_deadline, parse_iso
 
 __all__ = [
     "ascii_username",
+    "is_admin",
     "parse_predictions",
     "parse_line_predictions",
     "format_standings",
